+++ packages/cryptoplease/lib/app/screens/authenticated/wallet_flow/wallet_main_screen.dart
@@ -1,10 +1,10 @@
 import 'package:cryptoplease/app/components/number_formatter.dart';
-import 'package:cryptoplease/app/components/token_fiat_input_widget/enter_amount_keypad.dart';
 import 'package:cryptoplease/core/amount.dart';
 import 'package:cryptoplease/features/outgoing_direct_payments/presentation/odp_header.dart';
 import 'package:cryptoplease/gen/assets.gen.dart';
 import 'package:cryptoplease/l10n/device_locale.dart';
 import 'package:cryptoplease/l10n/l10n.dart';
+import 'package:cryptoplease/ui/amount_keypad/amount_keypad.dart';
 import 'package:cryptoplease/ui/button.dart';
 import 'package:cryptoplease/ui/info_icon.dart';
 import 'package:cryptoplease/ui/info_widget.dart';
@@ -99,7 +99,7 @@ class _ScreenState extends State<WalletMainScreen> {
           ),
           Flexible(
             child: LayoutBuilder(
-              builder: (context, constraints) => EnterAmountKeypad(
+              builder: (context, constraints) => AmountKeypad(
                 height: constraints.maxHeight,
                 width: width,
                 controller: _amountController,
