+++ packages/cryptoplease/lib/features/app_lock/presentation/components/pin_keypad.dart
@@ -1,4 +1,4 @@
-import 'package:cryptoplease/app/components/keypad_key.dart';
+import 'package:cryptoplease/ui/amount_keypad/keypad_key.dart';
 import 'package:flutter/material.dart';
 
 class PinKeypad extends StatelessWidget {
