+++ packages/cryptoplease/lib/features/payment_request/presentation/link_request/routes.dart
@@ -1,12 +1,10 @@
 import 'package:auto_route/auto_route.dart';
 import 'package:cryptoplease/features/payment_request/presentation/link_request/flow.dart';
 import 'package:cryptoplease/features/payment_request/presentation/link_request/payer_name_screen.dart';
-import 'package:cryptoplease/features/payment_request/presentation/link_request/request_amount_screen.dart';
 
 const linkRequestRoutes = AutoRoute<void>(
   page: LinkRequestFlowScreen,
   children: [
     AutoRoute<void>(path: '', page: PayerNameScreen),
-    AutoRoute<void>(page: RequestAmountScreen),
   ],
 );
